sdm660-common: Don't pin launcher app in memory
Similar to what we did for the camera app, unpin the launcher app from
memory as well. While the default launcher (Launcher3) isn't
particularly big, it doesn't make much sense to pin because the launcher
does not typically load new resources much. Most of its resources should
already be loaded in memory after it starts, so pinning the APK is
redundant.

NB: Default is false in fwb so remove it here

Change-Id: I6b82f2facb1bf4aee393af3f0bc3e0bce2538222
Signed-off-by: clarencelol <clarencekuiek@icloud.com>